Refactoring
- index.js 파일을 class component 별로 분할
- square.js, board.js, game.js
