+++ my-app/src/square.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+export class Square extends React.Component {
+    render() {
+      return (
+        <button className="square" onClick={() => this.props.onClick()}>
+          {this.props.value}
+        </button>
+      );
+    }
+  }
